requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework
gunicorn==21.2.0
packaging==23.1
Pillow==10.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
dj-database-url==0.5.0
whitenoise==5.3.0
psycopg2>=2.9,<3.0
djangorestframework-nested==0.1.0